+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add util methods: `makeHiddenInput()` and `appendHiddenInput()`.
+
 ### Fixed
-- Fix util method "toggleShow()".
+- Fix util method `toggleShow()`.
 
 ## [1.0.0] - 2017-10-07
 First release. Please see [ElementUtil API](https://github.com/archco/element-util/tree/master/docs).